+++ tomcat/src/main/java/org/apache/coyote/http11/auth/AuthService.java
@@ -3,10 +3,6 @@ package org.apache.coyote.http11.auth;
 import java.util.UUID;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
-import org.apache.coyote.http11.common.Cookie;
-import org.apache.coyote.http11.common.Session;
-import org.apache.coyote.http11.common.SessionRepository;
-import org.apache.coyote.http11.common.SessionRepositoryImpl;
 import org.apache.coyote.http11.request.body.RequestBody;
 import org.apache.coyote.http11.request.headers.RequestHeader;
 import org.apache.coyote.http11.request.line.HttpMethod;
