+++ tomcat/src/main/java/org/apache/coyote/http11/auth/Session.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.common;
+package org.apache.coyote.http11.auth;
 
 import java.util.HashMap;
 import java.util.Map;
