+++ tomcat/src/main/java/org/apache/coyote/http11/auth/SessionRepository.java
@@ -1,11 +1,9 @@
-package org.apache.coyote.http11.common;
+package org.apache.coyote.http11.auth;
 
 public interface SessionRepository {
 
     void create(final Session session);
 
     Session getSession(final String id);
-
-    void remove(final String id);
 
 }
