+++ tomcat/src/main/java/org/apache/coyote/http11/request/headers/RequestHeader.java
@@ -4,7 +4,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
-import org.apache.coyote.http11.common.Cookie;
+import org.apache.coyote.http11.auth.Cookie;
 
 import static java.util.Arrays.asList;
 import static java.util.Collections.EMPTY_LIST;
